+++ tests/backends/test_connection_params.py
@@ -1,5 +1,6 @@
+from unittest.mock import AsyncMock, patch
+
 import asyncpg
-from asynctest.mock import CoroutineMock, patch
 
 from tortoise import Tortoise
 from tortoise.contrib import test
@@ -7,7 +8,7 @@ from tortoise.contrib import test
 
 class TestConnectionParams(test.TestCase):
     async def test_mysql_connection_params(self):
-        with patch("asyncmy.create_pool", new=CoroutineMock()) as mysql_connect:
+        with patch("asyncmy.create_pool", new=AsyncMock()) as mysql_connect:
             await Tortoise._init_connections(
                 {
                     "models": {
@@ -42,7 +43,7 @@ class TestConnectionParams(test.TestCase):
 
     async def test_postres_connection_params(self):
         try:
-            with patch("asyncpg.create_pool", new=CoroutineMock()) as asyncpg_connect:
+            with patch("asyncpg.create_pool", new=AsyncMock()) as asyncpg_connect:
                 await Tortoise._init_connections(
                     {
                         "models": {