+++ tests/backends/test_mysql.py
@@ -15,7 +15,7 @@ class TestMySQL(test.SimpleTestCase):
         if self.db_config["connections"]["models"]["engine"] != "tortoise.backends.mysql":
             raise test.SkipTest("MySQL only")
 
-    async def tearDown(self) -> None:
+    async def asyncTearDown(self) -> None:
         if Tortoise._inited:
             await Tortoise._drop_databases()