+++ tests/contrib/test_pydantic.py
@@ -6,7 +6,7 @@ from tortoise.contrib.pydantic import pydantic_model_creator, pydantic_queryset_
 
 
 class TestPydantic(test.TestCase):
-    async def setUp(self) -> None:
+    async def asyncSetUp(self) -> None:
         self.Event_Pydantic = pydantic_model_creator(Event)
         self.Event_Pydantic_List = pydantic_queryset_creator(Event)
         self.Tournament_Pydantic = pydantic_model_creator(Tournament)
@@ -1035,7 +1035,7 @@ class TestPydantic(test.TestCase):
 
 
 class TestPydanticCycle(test.TestCase):
-    async def setUp(self) -> None:
+    async def asyncSetUp(self) -> None:
         self.Employee_Pydantic = pydantic_model_creator(Employee)
 
         self.root = await Employee.create(name="Root")