+++ tests/test_only.py
@@ -4,7 +4,7 @@ from tortoise.exceptions import IncompleteInstanceError
 
 
 class TestOnlyStraight(test.TestCase):
-    async def setUp(self) -> None:
+    async def asyncSetUp(self) -> None:
         self.model = StraightFields
         self.instance = await self.model.create(chars="Test")
 
@@ -59,6 +59,6 @@ class TestOnlyStraight(test.TestCase):
 
 
 class TestOnlySource(TestOnlyStraight):
-    async def setUp(self) -> None:
+    async def asyncSetUp(self) -> None:
         self.model = SourceFields  # type: ignore
         self.instance = await self.model.create(chars="Test")